[level 1] Title: 정수 제곱근 판별, Time: 0.05 ms, Memory: 75.6 MB -BaekjoonHub

+++ 프로그래머스/1/12934. 정수 제곱근 판별/정수 제곱근 판별.java
@@ -1,11 +1,11 @@
 
 class Solution {
-    public double solution(double n) {
-        double answer = -1;
+    public long solution(long n) {
+        long answer = -1;
         double result = Math.sqrt(n);
         
-        if (result == (int)result) {
-            answer = Math.pow(result + 1, 2);
+        if (result == (int) result) {
+            answer = (long) Math.pow(result + 1, 2);
         }
         
         return answer;
